+++ src/BlockBuilder/FieldType/Type/Number/Generation/NumberFieldGenerationContributor.php
@@ -14,6 +14,7 @@ use BlockBuilder\BlockGenerator\Generation\Plan\CodeFragment;
 use BlockBuilder\BlockGenerator\Generation\Plan\ControllerProperty;
 use BlockBuilder\BlockGenerator\Generation\Plan\DatabaseColumn;
 use BlockBuilder\BlockGenerator\Generation\Plan\Enum\ControllerMethodSectionEnum;
+use BlockBuilder\BlockGenerator\Generation\Plan\ViewVariableDocumentation;
 use BlockBuilder\FieldType\Enum\FieldTypeEnum;
 use BlockBuilder\FieldType\Type\Number\NumberFieldType;
 use BlockBuilder\FieldType\Type\Number\NumberFieldTypeDto;
@@ -63,6 +64,16 @@ readonly class NumberFieldGenerationContributor implements FieldGenerationContri
         } else {
             $this->contributeRepeatableControllerCode($field, $fragmentKeyPrefix, $context->position, $planBuilder);
         }
+
+        $planBuilder->view->addFieldVariable(
+            $context->fieldContext,
+            new ViewVariableDocumentation(
+                name: $field->handle,
+                type: 'int|float|string|null',
+                description: $field->label,
+                order: $context->position * 10,
+            ),
+        );
 
         $planBuilder->form->addFieldFragment(
             $context->fieldContext,